lnurl server: move zap receipt publishing to zap module
